created prop of Appointment
